Address review comments on tests
